temp

+++ src/main/java/subway/line/LineService.java
@@ -1,6 +1,7 @@
 package subway.line;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import subway.station.Station;
 import subway.station.StationRepository;
@@ -19,22 +20,18 @@ public final class LineService {
   }
 
   // 1. 지하철 노선 등록
-  public void addLine(
-      final String lineName,
-      final String firstStationName,
-      final String lastStationName
-  ) {
-    final Optional<Line> foundLine = lineRepository.findByLineName(lineName);
+  public void addLine(final CreateLine createLine) {
+    final Optional<Line> foundLine = lineRepository.findLineByName(createLine.getLineName());
+
     if (foundLine.isPresent()) {
       throw new IllegalArgumentException("중복된 노선명입니다");
     }
 
-    final Line line = createLine(lineName, firstStationName, lastStationName);
-    lineRepository.addLine(line);
+    lineRepository.addLine(createLine);
   }
 
   // 지하철 노선 생성
-  private Line createLine(
+  private void createLine(
       final String lineName,
       final String firstStationName,
       final String lastStationName
@@ -45,21 +42,30 @@ public final class LineService {
     if (firstStation.isEmpty() || lastStation.isEmpty()) {
       throw new IllegalArgumentException("유효하지 않은 상행 or 하행역명입니다.");
     }
-    return Line.of(lineName, firstStation.get(), lastStation.get());
+    CreateLine createLine = CreateLine.of(lineName, firstStation.get(), lastStation.get());
+    lineRepository.addLine(createLine);
   }
 
   // 2. 지하철 노선 삭제
   public void deleteLine(final String lineName) {
-    final Optional<Line> foundLine = lineRepository.findByLineName(lineName);
+    final Optional<Line> foundLine = lineRepository.findLineByName(lineName);
     if (foundLine.isEmpty()) {
       throw new IllegalArgumentException("존재하지 않는 노선입니다.");
     }
-    // TODO: 해당 노선이 포함하는 구간이 존재하는 경우 삭제 불가
+
+    final List<Station> stations = lineRepository.findStationsByLine(foundLine.get());
+    if (!stations.isEmpty()) {
+      throw new IllegalStateException("해당 노선이 포함하는 구간이 존재하므로 삭제 불가입니다");
+    }
     lineRepository.deleteLineByName(lineName);
   }
 
   // 3. 지하철 노선 조회
-  public List<Line> findAll() {
-    return lineRepository.findAll();
+  public Map<Line, List<Station>> getSubwayMap() {
+    return lineRepository.getSubwayMap();
+  }
+
+  public List<Line> findAllLines() {
+    return lineRepository.findAllLines();
   }
 }
